It works! Live channels are now integrated into the app, using the now independent ChannelSurfer library (Moreso independent)
