--- app/src/main/java/com/felkertech/channelsurfer/model/Stream.java
@@ -1,10 +1,0 @@
-package com.felkertech.channelsurfer.model;
-
-/**
- * Created by guest1 on 1/6/2016.
- */
-public class Stream {
-    public static final String VALUE_VIDEO_TYPE_HTTP_PROGRESSIVE = "HTTP_PROGRESSIVE";
-    public static final String VALUE_VIDEO_TYPE_HLS = "HLS";
-    public static final String VALUE_VIDEO_TYPE_MPEG_DASH = "MPEG_DASH";
-}
