--- app/src/main/java/com/felkertech/channelsurfer/service/SimpleTvSetup.java
@@ -1,7 +1,0 @@
-package com.felkertech.channelsurfer.service;
-
-/**
- * Created by guest1 on 1/6/2016.
- */
-public class SimpleTvSetup extends SetupTvInputProviderActivity {
-}
