+++ app/src/main/java/com/hitherejoe/vineyard/data/remote/VineyardService.java
@@ -51,6 +51,9 @@ public interface VineyardService {
     @GET("timelines/promoted")
     Observable<PostResponse> getEditorsPicksPosts(@Query("page") String page, @Query("anchorStr") String anchor);
 
+    @GET("timelines/promoted")
+    Call<PostResponse> getEditorsPicksPosts();
+
     @GET("timelines/tags/{tag}")
     Observable<PostResponse> getPostsByTag(@Path("tag") String tag, @Query("page") String page, @Query("anchorStr") String anchor);
