+++ app/src/main/java/com/hitherejoe/vineyard/ui/widget/LoopingVideoView.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.media.MediaPlayer;
 import android.net.Uri;
 import android.util.AttributeSet;
+import android.util.Log;
 import android.widget.VideoView;
 
 public class LoopingVideoView extends VideoView {
@@ -37,6 +38,7 @@ public class LoopingVideoView extends VideoView {
                 onVideoReadyListener.onVideoReady();
             }
         });
+        Log.d("LoopingVideoView", url);
         setVideoURI(Uri.parse(url));
     }
