+++ app/src/main/res/layout/activity_video_player.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <SurfaceView
+        android:layout_width="match_parent"
+        android:layout_marginTop="8dp"
+        android:id="@+id/surface"
+        android:layout_gravity="center_horizontal"
+        android:layout_height="match_parent" />
+</LinearLayout>
