--- app/src/main/res/layout/channel_surfer_setup.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    <TextView
-        android:layout_width="match_parent"
-        android:textSize="24sp"
-        android:text="Setting up channels..."
-        android:layout_height="wrap_content" />
-</LinearLayout>
