--- app/src/main/res/xml/channel_surfer_tv_input.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<tv-input xmlns:android="http://schemas.android.com/apk/res/android"
-    android:setupActivity="com.felkertech.channelsurfer.SimpleTvSetup"/>
