+++ channelsurfer/src/main/java/channelsurfer/service/SimpleTvSetup.java
@@ -1,0 +1,7 @@
+package channelsurfer.service;
+
+/**
+ * Created by guest1 on 1/6/2016.
+ */
+public class SimpleTvSetup extends SetupTvInputProviderActivity {
+}
