+++ channelsurfer/src/main/res/xml/channel_surfer_authenticator.xml
